+++ Specs/DataSources/ModelVisualizerSpec.js
@@ -21,6 +21,8 @@ import {
   ClippingPlane,
   ClippingPlaneCollection,
   Globe,
+  Cartographic,
+  createWorldTerrain,
 } from "../../Source/Cesium.js";
 import createScene from "../createScene.js";
 import pollToPromise from "../pollToPromise.js";
@@ -371,38 +373,150 @@ describe(
       });
     });
 
-    it("Computes bounding sphere with height reference", function () {
+    it("Computes bounding sphere with height reference clamp to ground", function () {
+      // Setup a position for the model.
+      const position = Cartesian3.fromDegrees(149.515332, -34.984799);
+      const positionCartographic = Cartographic.fromCartesian(position);
+
+      // Setup a spy so we can track how often sampleTerrain is called.
+      const sampleTerrainSpy = spyOn(
+        ModelVisualizer,
+        "_sampleTerrainMostDetailed"
+      ).and.callThrough();
+
+      // Initialize the Entity and the ModelGraphics.
       const time = JulianDate.now();
       const testObject = entityCollection.getOrCreateEntity("test");
       const model = new ModelGraphics({
         heightReference: HeightReference.CLAMP_TO_GROUND,
       });
       testObject.model = model;
-
-      testObject.position = new ConstantProperty(
-        new Cartesian3(5678, 1234, 1101112)
-      );
+      testObject.position = new ConstantProperty(position);
       model.uri = new ConstantProperty(boxUrl);
+
       visualizer.update(time);
 
-      const modelPrimitive = scene.primitives.get(0);
+      // Request the bounding sphere once.
       const result = new BoundingSphere();
       let state = visualizer.getBoundingSphere(testObject, result);
       expect(state).toBe(BoundingSphereState.PENDING);
 
-      return pollToPromise(function () {
-        scene.render();
-        state = visualizer.getBoundingSphere(testObject, result);
-        return state !== BoundingSphereState.PENDING;
-      }).then(function () {
-        expect(state).toBe(BoundingSphereState.DONE);
-        expect(modelPrimitive._clampedModelMatrix).toBeDefined();
-        const expected = BoundingSphere.clone(
-          modelPrimitive.boundingSphere,
-          new BoundingSphere()
-        );
-        expect(result).toEqual(expected);
+      // Assign a tiled terrain provider to the globe.
+      const globe = scene.globe;
+      const previousTerrainProvider = globe.terrainProvider;
+      globe.terrainProvider = createWorldTerrain();
+
+      let sampledResultCartographic;
+      let sampledResult;
+
+      return ModelVisualizer._sampleTerrainMostDetailed(globe.terrainProvider, [
+        positionCartographic,
+      ])
+        .then((updatedCartographics) => {
+          sampledResultCartographic = updatedCartographics[0];
+          sampledResult = globe.ellipsoid.cartographicToCartesian(
+            sampledResultCartographic
+          );
+
+          // Repeatedly request the bounding sphere until it's ready.
+          return pollToPromise(function () {
+            scene.render();
+            state = visualizer.getBoundingSphere(testObject, result);
+            return state !== BoundingSphereState.PENDING;
+          });
+        })
+        .then(() => {
+          expect(state).toBe(BoundingSphereState.DONE);
+
+          // Ensure that we only sample the terrain once from the visualizer.
+          // We check for 2 calls here because we call it once in the test.
+          expect(sampleTerrainSpy).toHaveBeenCalledTimes(2);
+
+          // Calculate the distance of the bounding sphere returned from the position returned from sample terrain.
+          // Since sampleTerrainMostDetailed isn't always precise, we account for some error.
+          const distance = Cartesian3.distance(result.center, sampledResult);
+          const errorMargin = 100.0;
+          expect(distance).toBeLessThan(errorMargin);
+
+          // Reset the terrain provider.
+          globe.terrainProvider = previousTerrainProvider;
+        });
+    });
+
+    it("Computes bounding sphere with height reference relative to ground", function () {
+      // Setup a position for the model.
+      const heightOffset = 1000.0;
+      const position = Cartesian3.fromDegrees(
+        149.515332,
+        -34.984799,
+        heightOffset
+      );
+      const positionCartographic = Cartographic.fromCartesian(position);
+
+      // Setup a spy so we can track how often sampleTerrain is called.
+      const sampleTerrainSpy = spyOn(
+        ModelVisualizer,
+        "_sampleTerrainMostDetailed"
+      ).and.callThrough();
+
+      // Initialize the Entity and the ModelGraphics.
+      const time = JulianDate.now();
+      const testObject = entityCollection.getOrCreateEntity("test");
+      const model = new ModelGraphics({
+        heightReference: HeightReference.RELATIVE_TO_GROUND,
       });
+      testObject.model = model;
+      testObject.position = new ConstantProperty(position);
+      model.uri = new ConstantProperty(boxUrl);
+
+      visualizer.update(time);
+
+      // Request the bounding sphere once.
+      const result = new BoundingSphere();
+      let state = visualizer.getBoundingSphere(testObject, result);
+      expect(state).toBe(BoundingSphereState.PENDING);
+
+      // Assign a tiled terrain provider to the globe.
+      const globe = scene.globe;
+      const previousTerrainProvider = globe.terrainProvider;
+      globe.terrainProvider = createWorldTerrain();
+
+      let sampledResultCartographic;
+      let sampledResult;
+
+      return ModelVisualizer._sampleTerrainMostDetailed(globe.terrainProvider, [
+        positionCartographic,
+      ])
+        .then((updatedCartographics) => {
+          sampledResultCartographic = updatedCartographics[0];
+          sampledResult = globe.ellipsoid.cartographicToCartesian(
+            sampledResultCartographic
+          );
+
+          // Repeatedly request the bounding sphere until it's ready.
+          return pollToPromise(function () {
+            scene.render();
+            state = visualizer.getBoundingSphere(testObject, result);
+            return state !== BoundingSphereState.PENDING;
+          });
+        })
+        .then(() => {
+          expect(state).toBe(BoundingSphereState.DONE);
+
+          // Ensure that we only sample the terrain once from the visualizer.
+          // We check for 2 calls here because we call it once in the test.
+          expect(sampleTerrainSpy).toHaveBeenCalledTimes(2);
+
+          // Calculate the distance of the bounding sphere returned from the position returned from sample terrain.
+          // Since sampleTerrainMostDetailed isn't always precise, we account for some error.
+          const distance =
+            Cartesian3.distance(result.center, sampledResult) - heightOffset;
+          const errorMargin = 100.0;
+          expect(distance).toBeLessThan(errorMargin);
+
+          // Reset the terrain provider.
+          globe.terrainProvider = previousTerrainProvider;
+        });
     });
 
     it("Fails bounding sphere for entity without billboard.", function () {